pyproject.toml
[project]
name = "deep-researcher-openai-sdk"
version = "0.2.0"
description = "Modular AI Research Framework with OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Adrian Marquez", email = "adrianmf2194@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gradio>=5.22.0",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.0.1",
    "sendgrid>=6.11.0",
    "pydantic>=2.0.0",
]

[project.scripts]
deep-research-cruise = "deep_researcher.core.domain_launcher:main"
deep-research-job = "deep_researcher.core.domain_launcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deep_researcher"]